ратинг в форме вынес в отделный компонент +в обрабочик передаю i
